moved wordcloud_functions to utils folder, reinstated all main repo endpoints
