+++ app/main.py
@@ -1,9 +1,15 @@
-from fastapi import FastAPI
-from fastapi.middleware.cors import CORSMiddleware
-import uvicorn
-from app.api import wordcloud_text, wordcloud_cloud, wordcloud_update
+from os import getenv
 
-application = app = FastAPI(
+from fastapi import FastAPI, Security
+from fastapi.middleware.cors import CORSMiddleware
+from fastapi.responses import Response, JSONResponse
+from fastapi.requests import Request
+import uvicorn
+
+from app.api import submission, visualization, clustering, db, wordcloud_text, wordcloud_cloud, wordcloud_update
+from app.utils.security.header_checking import get_api_key
+
+app = FastAPI(
     title="Labs26-StorySquad-DS-Team B",
     description="A RESTful API for the Story Squad Project",
     version="0.1",
@@ -13,6 +19,10 @@ application = app = FastAPI(
 app.include_router(wordcloud_text.router, tags=['Word Cloud'])
 app.include_router(wordcloud_update.router, tags=['Word Cloud'])
 app.include_router(wordcloud_cloud.router, tags=['Word Cloud'])
+app.include_router(submission.router, tags=['Submission'], dependencies=[Security(get_api_key)])
+app.include_router(visualization.router, tags=['Visualization'], dependencies=[Security(get_api_key)])
+app.include_router(clustering.router, tags=['Clustering'], dependencies=[Security(get_api_key)])
+app.include_router(db.router, tags=['Database'], dependencies=[Security(get_api_key)])
 
 app.add_middleware(
     CORSMiddleware,
@@ -23,4 +33,4 @@ app.add_middleware(
 )
 
 if __name__ == "__main__":
-    uvicorn.run(application)
+    uvicorn.run(app)
